chore(module:descriptions): refactor
