--- components/descriptions/nz-descriptions-item.component.html
@@ -1,4 +1,0 @@
-<!-- Use a template to wrap contents so contents wouldn't be displayed. -->
-<ng-template>
-  <ng-content></ng-content>
-</ng-template>
