+++ components/descriptions/typings.ts
